requirements.txt
anyio==3.6.2
base58==2.1.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pycodestyle==2.10.0
pyTelegramBotAPI==4.10.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
types-Deprecated==1.2.9.1
typing_extensions==4.5.0
urllib3==1.26.14
websockets==10.4
wrapt==1.15.0
xrpl-py==1.7.0
